chore: move react emails into separate directories
