+++ app/emails/private-cloud/AdminRequest.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import NewRequestTemplate from './_templates/private-cloud/CreateRequest';
+import { samplePrivateRequest } from '../_components/Params';
+import NewRequestTemplate from '../_templates/private-cloud/AdminEditRequest';
 
 export default function NewRequest() {
   return <NewRequestTemplate request={samplePrivateRequest} userName={'Session User'} />;
