+++ app/emails/private-cloud/CreateRequest.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import NewRequestTemplate from './_templates/private-cloud/AdminDeleteRequest';
+import { samplePrivateRequest } from '../_components/Params';
+import NewRequestTemplate from '../_templates/private-cloud/CreateRequest';
 
 export default function NewRequest() {
   return <NewRequestTemplate request={samplePrivateRequest} userName={'Session User'} />;
