+++ app/emails/private-cloud/CreateRequestApproval.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import CreateRequestApprovalTemplate from './_templates/private-cloud/CreateRequestApproval';
+import { samplePrivateRequest } from '../_components/Params';
+import CreateRequestApprovalTemplate from '../_templates/private-cloud/CreateRequestApproval';
 
 export default function CreateRequestApproval() {
   return <CreateRequestApprovalTemplate request={samplePrivateRequest} />;
