+++ app/emails/private-cloud/DeleteCompletion.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateProduct } from './_components/Params';
-import DeleteApprovalTemplate from './_templates/private-cloud/DeleteApproval';
+import { samplePrivateProduct } from '../_components/Params';
+import DeleteApprovalTemplate from '../_templates/private-cloud/DeleteApproval';
 
 export default function DeleteApproval() {
   return <DeleteApprovalTemplate product={samplePrivateProduct} />;
