+++ app/emails/private-cloud/DeleteRequest.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import DeleteTemplate from './_templates/private-cloud/DeleteRequest';
+import { samplePrivateRequest } from '../_components/Params';
+import DeleteTemplate from '../_templates/private-cloud/DeleteRequest';
 
 export default function DeleteRequest() {
   return <DeleteTemplate request={samplePrivateRequest} userName={'User Session'} />;
