+++ app/emails/private-cloud/DeleteRequestApproval.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import DeleteRequestApprovalTemplate from './_templates/private-cloud/DeleteRequestApproval';
+import { samplePrivateRequest } from '../_components/Params';
+import DeleteRequestApprovalTemplate from '../_templates/private-cloud/DeleteRequestApproval';
 
 export default function DeleteRequestApproval() {
   return <DeleteRequestApprovalTemplate request={samplePrivateRequest} />;
