+++ app/emails/private-cloud/EditRequestApproval.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import EditRequestApprovalTemplate from './_templates/private-cloud/EditRequestApproval';
+import { samplePrivateRequest } from '../_components/Params';
+import EditRequestApprovalTemplate from '../_templates/private-cloud/EditRequestApproval';
 
 export default function EditRequestApproval() {
   return <EditRequestApprovalTemplate request={samplePrivateRequest} />;
