+++ app/emails/private-cloud/EditRequestCompletion.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateProduct } from './_components/Params';
-import EditRequestCompleteTemplate from './_templates/private-cloud/EditRequestComplete';
+import { samplePrivateProduct } from '../_components/Params';
+import EditRequestCompleteTemplate from '../_templates/private-cloud/EditRequestComplete';
 
 export default function EditRequestComplete() {
   return <EditRequestCompleteTemplate product={samplePrivateProduct} />;
