+++ app/emails/private-cloud/RequestProvisioned.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateProduct } from './_components/Params';
-import ProvisionedTemplate from './_templates/private-cloud/Provisioned';
+import { samplePrivateProduct } from '../_components/Params';
+import ProvisionedTemplate from '../_templates/private-cloud/Provisioned';
 
 export default function EditRequest() {
   return <ProvisionedTemplate product={samplePrivateProduct} />;
