+++ app/emails/private-cloud/RequestRejection.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePrivateRequest } from './_components/Params';
-import RequestRejectionTemplate from './_templates/private-cloud/RequestRejection';
+import { samplePrivateRequest } from '../_components/Params';
+import RequestRejectionTemplate from '../_templates/private-cloud/RequestRejection';
 
 export default function RequestRejection() {
   return <RequestRejectionTemplate request={samplePrivateRequest} currentData={samplePrivateRequest.decisionData} />;
