+++ app/emails/public-cloud/AdminDeleteRequests.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicRequest } from './_components/Params';
-import NewDeleteRequestTemplate from './_templates/public-cloud/AdminDeleteRequest';
+import { samplePublicRequest } from '../_components/Params';
+import NewDeleteRequestTemplate from '../_templates/public-cloud/AdminDeleteRequest';
 
 export default function NewRequest() {
   // Assuming samplePublicRequest has a decisionData property of the correct type
