+++ app/emails/public-cloud/DeleteRequest.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicProduct } from './_components/Params';
-import DeleteTemplate from './_templates/public-cloud/DeleteRequest';
+import { samplePublicProduct } from '../_components/Params';
+import DeleteTemplate from '../_templates/public-cloud/DeleteRequest';
 
 export default function DeleteRequest() {
   return <DeleteTemplate product={samplePublicProduct} userName={'User Session'} />;
