+++ app/emails/public-cloud/ExpenseAuthority.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicProduct } from './_components/Params';
-import ExpenseAuthority from './_templates/public-cloud/ExpenseAuthority';
+import { samplePublicProduct } from '../_components/Params';
+import ExpenseAuthority from '../_templates/public-cloud/ExpenseAuthority';
 
 export default function PublicCloudExpenseAuthority() {
   return <ExpenseAuthority product={samplePublicProduct} />;
