+++ app/emails/public-cloud/ExpenseAuthorityMou.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicRequest } from './_components/Params';
-import ExpenseAuthorityMou from './_templates/public-cloud/ExpenseAuthorityMou';
+import { samplePublicRequest } from '../_components/Params';
+import ExpenseAuthorityMou from '../_templates/public-cloud/ExpenseAuthorityMou';
 
 export default function PublicCloudExpenseAuthorityMou() {
   return <ExpenseAuthorityMou request={samplePublicRequest} />;
