+++ app/emails/public-cloud/RequestApproval.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicRequest } from './_components/Params';
-import RequestApprovalTemplate from './_templates/public-cloud/RequestApproval';
+import { samplePublicRequest } from '../_components/Params';
+import RequestApprovalTemplate from '../_templates/public-cloud/RequestApproval';
 
 export default function RequestApproval() {
   return <RequestApprovalTemplate request={samplePublicRequest} />;
