+++ app/emails/public-cloud/RequestRejection.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { samplePublicRequest } from './_components/Params';
-import RequestRejectionTemplate from './_templates/public-cloud/RequestRejection';
+import { samplePublicRequest } from '../_components/Params';
+import RequestRejectionTemplate from '../_templates/public-cloud/RequestRejection';
 
 export default function RequestRejection() {
   // Extract the product name and human comment from the samplePublicRequest
